requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
pandas>=2.0.0
pyarrow>=14.0.0
pyyaml>=6.0
rich>=13.0.0
numpy>=1.24.0
scipy>=1.11.0
